Transfer relevant tests to the correct test file
